requirements.txt
matplotlib==3.1.1
nltk==3.4.5
numpy==1.17.1
pandas==0.25.1
Pillow==6.2.2
pyparsing==2.4.2
python-dateutil==2.8.0
py-opencv=4.5.3
pyyaml
requests==2.25.1
rtpt
seaborn
scikit-image
scikit-learn
scipy
torch==1.8.0
torchattacks
torchmetrics
torchvision==0.9.0
torchsummary
timm
